build(webpack): reinstall
After cloning the project and spinning up the server visiting the
`localhost:300` I got a
`Webpacker::Manifest::MissingEntryError in Pages#welcome`.
Knowing that the webpacker version was changed and the aforementioned
error can be caused by a misconfigured webpacker I run a
`rake webpacker:install`.
The only thing to tune after the fresh install was to add back the
`tailwind` requirement at `postcss.confifg.js`.

**Important**: This commit also includes changes in the schema product
of running the migrations. Maybe a previous conflict left a corrupted
version.